chore: move LoRA related components into own folder
