+++ app/_components/AdvancedOptions/LoRAs/LoraSettingsCard.tsx
@@ -1,11 +1,11 @@
 import { SavedLora } from '@/app/_types/ArtbotTypes'
-import Button from '../Button'
+import Button from '../../Button'
 import { IconInfoCircle, IconTrash } from '@tabler/icons-react'
 import { useInput } from '@/app/_providers/PromptInputProvider'
-import NumberInput from '../NumberInput'
-import OptionLabel from './OptionLabel'
+import NumberInput from '../../NumberInput'
+import OptionLabel from '../OptionLabel'
 import { useState } from 'react'
-import DeleteConfirmation from '../Modal_DeleteConfirmation'
+import DeleteConfirmation from '../../Modal_DeleteConfirmation'
 import NiceModal from '@ebay/nice-modal-react'
 import LoraDetails from './LoraDetails'
